requirements.txt
markdown
pyserial
protobuf
dotmap
pexpect
pyqrcode
tabulate
timeago
webencodings
pyparsing
twine
autopep8
pylint
pytest
pytest-cov
pyyaml
pytap2
pdoc3
pypubsub
pygatt; platform_system == "Linux"
geopy==2.3.0
PyNaCl==1.5.0
